Adds error handling in networkRequest
Sometimes the requets fails randomly and/or the content type is not json.
The lack of error handling makes consumers unable to catch the error to see what went wrong.

This commit adds error handling for both cases when the request fails (status !== 2xx)
and when the content-type is not json.